chore: refactor for consistent code layout
have all api packages/modules use the same folder structure
