+++ backend/src/ches/v1/core/schemas/ches-message-object.ts
