+++ backend/src/ches/v1/core/schemas/ches-status-object.ts
