+++ backend/src/defaults/v1/core/schemas/index.ts
@@ -1,0 +1,1 @@
+export * from './tenant-defaults';
